D3CF_boost/piprequirement.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall==0.2.0
bleach==4.1.0
blis==0.7.5
boto==2.49.0
boto3 @ file:///tmp/build/80754af9/boto3_1629142476879/work
botocore @ file:///tmp/build/80754af9/botocore_1633531073272/work
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
ConfigArgParse==1.5.3
crcmod==1.7
cryptography @ file:///tmp/build/80754af9/cryptography_1639400846433/work
cycler==0.11.0
cymem==2.0.6
cytoolz==0.11.2
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
diffdist==0.1
dill==0.3.4
entrypoints==0.3
faiss-gpu==1.6.5
fasteners==0.17.3
filelock==3.4.2
fire==0.4.0
flatbuffers==2.0
fonttools==4.28.5
frozenlist==1.3.0
fsspec==2022.11.0
gast==0.5.3
gcs-oauth2-boto-plugin==3.0
gensim==3.8.1
google-apitools==0.5.32
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-reauth==0.1.1
googledrivedownloader==0.4
grpcio==1.44.0
gsutil==5.8
gym==0.23.1
gym-notices==0.0.6
h5py==2.10.0
hickle==4.0.4
higher==0.2.1
httplib2==0.20.4
huggingface-hub==0.2.1
hydra-core==1.2.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.11.2
importlib-resources==5.4.0
install==1.3.5
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
jedi==0.18.1
Jinja2==3.0.3
jmespath @ file:///Users/ktietz/demo/mc3/conda-bld/jmespath_1630583964805/work
joblib==1.1.0
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langcodes==3.3.0
learn2learn==0.1.7
libclang==13.0.0
lightly==1.2.35
lightly-utils==0.0.2
lightning-utilities==0.3.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
monotonic==1.6
multidict==6.0.2
munkres==1.1.4
murmurhash==1.0.6
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.6.7
notebook==6.4.6
numexpr @ file:///tmp/build/80754af9/numexpr_1618856167419/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095647912/work
oauth2client==4.1.3
oauthlib==3.2.0
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
omegaconf==2.2.3
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.8.2
Pygments==2.11.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-lightning==1.8.1
pytorch-lightning-bolts==0.3.2.post1
pytorch-transformers==1.2.0
pytz==2021.3
pyu2f==0.1.5
PyYAML==6.0
pyzmq==22.3.0
qpth==0.0.15
qtconsole==5.2.2
QtPy==2.0.0
rdflib==6.1.1
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.1
retry-decorator==1.1.1
rsa==4.7.2
s3transfer @ file:///tmp/build/80754af9/s3transfer_1626435152308/work
sacremoses==0.0.46
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.96
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.10.3
toolz==0.11.2
torch==1.10.1
torch-cluster @ file:///home/hlx/Downloads/torch_cluster-1.5.9-cp38-cp38-linux_x86_64.whl
torch-geometric==2.0.3
torch-scatter @ file:///home/hlx/Downloads/torch_scatter-2.0.9-cp38-cp38-linux_x86_64.whl
torch-sparse @ file:///home/hlx/Downloads/torch_sparse-0.6.12-cp38-cp38-linux_x86_64.whl
torch-xla==1.0
torchaudio==0.10.1
torchmetrics==0.10.2
torchsummary==1.5.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.15.0
typer==0.4.0
typing_extensions==4.4.0
urllib3==1.26.7
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wrapt==1.13.3
yacs==0.1.8
yarl==1.7.2
zipp==3.6.0
